feat(frontend): render markdown in chat message bubbles
Extract MarkdownBody into shared component and use it in both the
chat content path and the MessagePresentationView summaryText,
so assistant replies with tables, headings, lists, and blockquotes
render properly inside conversation bubbles.
